test: Exclude SeedStartup from coverage counts
