+++ src/test/fuzz/fuzz.cpp
@@ -7,6 +7,7 @@
 #include <netaddress.h>
 #include <netbase.h>
 #include <test/fuzz/util/check_globals.h>
+#include <test/util/coverage.h>
 #include <test/util/random.h>
 #include <test/util/setup_common.h>
 #include <util/check.h>
@@ -88,25 +89,6 @@ static void test_one_input(FuzzBufferType buffer)
 const std::function<std::string()> G_TEST_GET_FULL_NAME{[]{
     return std::string{g_fuzz_target};
 }};
-
-#if defined(__clang__) && defined(__linux__)
-extern "C" void __llvm_profile_reset_counters(void) __attribute__((weak));
-extern "C" void __gcov_reset(void) __attribute__((weak));
-
-void ResetCoverageCounters()
-{
-    if (__llvm_profile_reset_counters) {
-        __llvm_profile_reset_counters();
-    }
-
-    if (__gcov_reset) {
-        __gcov_reset();
-    }
-}
-#else
-void ResetCoverageCounters() {}
-#endif
-
 
 static void initialize()
 {
