+++ src/test/util/coverage.h
@@ -1,0 +1,10 @@
+// Copyright (c) 2025-present The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_TEST_UTIL_COVERAGE_H
+#define BITCOIN_TEST_UTIL_COVERAGE_H
+
+void ResetCoverageCounters();
+
+#endif // BITCOIN_TEST_UTIL_COVERAGE_H
